1.所有“抽奖”字段+注释+接口名称对应前端变更为“召唤”。 2.系统封装通用策略模式；（目前已在召唤系统应用，后面其他场景也会更新）。 3.大幅优化提升召唤系统处理效率。 4.召唤系统结构整体更新升级。 5.优化部分数据库表字段格式以节省占用空间。 6.各模块字段命名规范化。 7.优化英雄技能表结构。 8.优化英雄升星功能处理效率。 9.弃用英雄星级奖池模块。 10.清理系统部分冗余代码。

+++ gm-domain/src/main/java/com/gm/modules/basicconfig/dto/AttributeEntity.java
@@ -18,7 +18,7 @@ public class AttributeEntity {
     /**
      * 英雄星级
      */
-    private Long heroStar;
+    private Integer heroStar;
     /**
      * 技能名称
      */
